Make PurposeCategoryList a container
This makes PurposeCategoryList more modular and makes it
easier to add links to the individual purposes.
Refs #29.
