+++ app/containers/PurposeCategoryList.js
@@ -1,13 +1,21 @@
 import _ from 'lodash';
 import React, { Component, PropTypes } from 'react';
 import Loader from 'react-loader';
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
 
+import { fetchPurposes } from 'actions/purposeActions';
 import PurposeCategory from 'components/purpose/PurposeCategory';
+import { purposeCategoryListSelectors } from 'selectors/purposeCategoryListSelectors';
 
-export class PurposeCategoryList extends Component {
+export class UnconnectedPurposeCategoryList extends Component {
   constructor(props) {
     super(props);
     this.renderPurposeCategory = this.renderPurposeCategory.bind(this);
+  }
+
+  componentDidMount() {
+    this.props.actions.fetchPurposes();
   }
 
   renderPurposeCategory(mainType) {
@@ -35,8 +43,13 @@ export class PurposeCategoryList extends Component {
   }
 }
 
-PurposeCategoryList.propTypes = {
+UnconnectedPurposeCategoryList.propTypes = {
+  actions: PropTypes.object.isRequired,
   purposeCategories: PropTypes.object.isRequired,
 };
 
-export default PurposeCategoryList;
+function mapDispatchToProps(dispatch) {
+  return { actions: bindActionCreators({ fetchPurposes }, dispatch) };
+}
+
+export default connect(purposeCategoryListSelectors, mapDispatchToProps)(UnconnectedPurposeCategoryList);
